legacy_python/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
chardet==3.0.4
discord.py==1.5.0
multidict==4.6.1
pkg-resources==0.0.0
websockets==8.1
